fz.orvibo_raw_1 and fz.orvibo_raw_2
